Arguehub-Backend/pyproject.toml
[project]
name = "arguehub-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.1",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "psycopg2>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "redis>=6.0.0",
    "rq>=2.3.2",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]>=0.34.2",
]
[project.scripts]
server = "api.v1.endpoints.server:main"
